databuilder/requirements.txt
feedparser==6.0.11
py-dateutil==2.2
PyYAML==6.0.2
sgmllib3k==1.0.0
six==1.16.0
